+++ qsprpred/models/hyperparam_optimization.py
@@ -11,7 +11,7 @@ from sklearn.model_selection import ParameterGrid
 from qsprpred.models.assessment.methods import ModelAssessor
 from ..data import QSPRDataset
 from ..logs import logger
-from ..models.models import QSPRModel
+from ..models.model import QSPRModel
 from ..models.monitors import BaseMonitor, HyperparameterOptimizationMonitor