+++ qsprpred/models/monitors.py
@@ -9,7 +9,7 @@ import pandas as pd
 from rdkit import Chem
 from rdkit.Chem import Draw
 
-from .models import QSPRModel
+from .model import QSPRModel
 from ..data.tables.qspr import QSPRDataset
 from ..utils.serialization import JSONSerializable